requirements.txt
fastapi[all]
uvicorn
python-jose[cryptography]
passlib[bcrypt]
